+++ libdariadb/storage/page.cpp
@@ -390,7 +390,7 @@ dariadb::Id2Meas Page::valuesBeforeTimePoint(const QueryTimePoint &q) {
     auto reader = c->get_reader();
     while (!reader->is_end()) {
       auto m = reader->readNext();
-      if (m.time <= q.time_point && m.inQuery(q.ids, q.flag, q.source)) {
+      if (m.time <= q.time_point && m.inQuery(q.ids, q.flag)) {
         auto f_res = result.find(m.id);
         if (f_res == result.end()) {
           to_read.erase(m.id);
@@ -441,7 +441,7 @@ void Page::readLinks(const QueryInterval &query, const ChunkLinkList &links,
       if (subres.time > query.to) {
         break;
       }
-      if (subres.inQuery(query.ids, query.flag, query.source, query.from, query.to)) {
+      if (subres.inQuery(query.ids, query.flag, query.from, query.to)) {
         clb->call(subres);
       }
     }